숙제1을 Homework1 클래스로 분리
